backend/pyproject.toml
[project]
name = "git-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.1",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "openai>=1.75.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.5",
    "pygithub>=2.9.0",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn[standard]>=0.34.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.11",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.pyright]
exclude = [".venv"]
venvPath = "."
venv = ".venv"
extraPaths = ["./fastapi-application"]

[tool.pytest.ini_options]
pythonpath = "fastapi-application"
testpaths = [
    "tests"
]
